# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Hot AI Backend API
  description: |
    Hot AI 后端服务 API 接口文档
    
    ## 认证说明
    除了注册和登录接口外，其他接口都需要在请求头中携带 JWT Token：
    ```
    Authorization: Bearer <your_access_token>
    ```
    
    Token 有效期为 24 小时，过期后可以使用 Refresh Token 刷新。
  version: 1.0.0
  contact:
    name: API Support
  license:
    name: MIT

servers:
  - url: http://localhost:8080/api/v1
    description: 本地开发环境
  - url: https://api.hot-ai.com/api/v1
    description: 生产环境

tags:
  - name: 认证
    description: 用户认证相关接口

paths:
  /auth/register:
    post:
      tags:
        - 认证
      summary: 用户注册
      description: |-
        创建新用户账号
        
        ## 注册流程说明
        1. 先调用 `/auth/send-registration-code` 发送邮箱验证码
        2. 用户收到验证码后，调用此接口提交注册信息
        3. 验证码有效期为 5 分钟，且只能使用一次
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              email: user@example.com
              password: SecurePass123!
              password_confirm: SecurePass123!
              nickname: 用户名
              verification_code: "123456"
      responses:
        '200':
          description: 注册成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
              example:
                user_id: user-123
                message: 注册成功，请登录
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponse'
              example:
                code: 400
                message: 邮箱格式不正确
                data: null

  /auth/login:
    post:
      tags:
        - 认证
      summary: 用户登录
      description: 用户登录获取访问令牌
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: user@example.com
              password: SecurePass123!
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              example:
                access_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                refresh_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                expires_in: 86400
                user:
                  id: user-123
                  email: user@example.com
                  nickname: 测试用户
                  avatar: ""
                  roles:
                    - user
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponse'
              example:
                code: 400
                message: 邮箱或密码格式错误
                data: null
        '401':
          description: 认证失败（邮箱或密码错误）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponse'
              example:
                code: 401
                message: 邮箱或密码错误
                data: null

  /auth/refresh:
    post:
      tags:
        - 认证
      summary: 刷新 Token
      description: 使用 Refresh Token 获取新的 Access Token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
            example:
              refresh_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      responses:
        '200':
          description: 刷新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshTokenResponse'
              example:
                access_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                expires_in: 86400
        '401':
          description: Refresh Token 无效或已过期
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponse'
              example:
                code: 401
                message: Refresh Token 无效或已过期
                data: null

  /auth/logout:
    post:
      tags:
        - 认证
      summary: 用户登出
      description: 退出登录，使当前 Token 失效
      operationId: logoutUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 登出成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponse'
              example:
                code: 200
                message: 登出成功
                data:
                  success: true
        '401':
          description: 未授权
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponse'
              example:
                code: 401
                message: 未授权访问
                data: null

  /auth/send-registration-code:
    post:
      tags:
        - 认证
      summary: 发送注册验证码
      description: |-
        向指定邮箱发送注册验证码，用于验证邮箱所有权
        
        ## 限流规则
        - 同一邮箱 1 分钟内最多发送 1 次
        - 同一邮箱 1 小时内最多发送 5 次
        
        ## 验证码说明
        - 验证码为 6 位数字
        - 有效期 5 分钟
        - 一次性使用，验证后自动失效
      operationId: sendRegistrationCode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                  description: 需要验证的邮箱地址
                  example: user@example.com
            example:
              email: user@example.com
      responses:
        '200':
          description: 验证码发送成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 验证码已发送，请检查邮箱
                  email:
                    type: string
                    example: user@example.com
                  expires_in:
                    type: integer
                    description: 验证码有效期（秒）
                    example: 300
        '400':
          description: 请求参数错误或邮箱已注册
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponse'
              example:
                code: 400
                message: 该邮箱已注册
                data: null
        '429':
          description: 操作过于频繁
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponse'
              example:
                code: 429
                message: 操作过于频繁，请稍后再试
                data: null

  /auth/send-verification-email:
    post:
      tags:
        - 认证
      summary: 发送验证邮件
      description: 向指定邮箱发送验证邮件，用于验证邮箱所有权
      operationId: sendVerificationEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                  description: 需要验证的邮箱地址
                  example: user@example.com
            example:
              email: user@example.com
      responses:
        '200':
          description: 邮件发送成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 验证邮件已发送，请检查邮箱
                  email:
                    type: string
                    example: user@example.com
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponse'
              example:
                code: 400
                message: 邮箱格式不正确
                data: null
        '429':
          description: 请求过于频繁（限流）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponse'
              example:
                code: 429
                message: 操作过于频繁，请稍后再试
                data: null

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer Token 认证

  schemas:
    # ============ 请求模型 ============
    RegisterRequest:
      type: object
      description: 用户注册请求
      required:
        - email
        - password
        - nickname
      properties:
        email:
          type: string
          format: email
          description: 邮箱地址
          example: user@example.com
        password:
          type: string
          format: password
          description: 密码
          minLength: 8
          maxLength: 32
          example: SecurePass123!
        nickname:
          type: string
          description: 用户昵称
          minLength: 2
          maxLength: 20
          example: 用户名

    LoginRequest:
      type: object
      description: 用户登录请求
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: 邮箱地址
          example: user@example.com
        password:
          type: string
          format: password
          description: 密码
          example: SecurePass123!

    RefreshTokenRequest:
      type: object
      description: 刷新 Token 请求
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
          description: Refresh Token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...

    # ============ 响应模型 ============
    RegisterResponse:
      type: object
      description: 注册响应
      properties:
        user_id:
          type: string
          description: 用户 ID
          example: user-123
        message:
          type: string
          description: 消息
          example: 注册成功，请登录

    LoginResponse:
      type: object
      description: 登录响应
      properties:
        access_token:
          type: string
          description: Access Token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refresh_token:
          type: string
          description: Refresh Token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        expires_in:
          type: integer
          description: 过期时间（秒）
          example: 86400
        user:
          $ref: '#/components/schemas/UserInfo'

    RefreshTokenResponse:
      type: object
      description: 刷新 Token 响应
      properties:
        access_token:
          type: string
          description: Access Token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        expires_in:
          type: integer
          description: 过期时间（秒）
          example: 86400

    CommonResponse:
      type: object
      description: 通用响应结构
      properties:
        code:
          type: integer
          description: 状态码
          example: 200
        message:
          type: string
          description: 消息
          example: 操作成功
        data:
          type: object
          description: 数据
          nullable: true

    UserInfo:
      type: object
      description: 用户基本信息
      properties:
        id:
          type: string
          description: 用户 ID
          example: user-123
        email:
          type: string
          format: email
          description: 邮箱
          example: user@example.com
        nickname:
          type: string
          description: 昵称
          example: 测试用户
        avatar:
          type: string
          description: 头像 URL
          example: ""
        roles:
          type: array
          description: 角色列表
          items:
            type: string
          example:
            - user

# ============ 全局错误响应 ============
x-error-codes:
  200: 请求成功
  400: 请求参数错误
  401: 未授权或认证失败
  403: 权限不足
  404: 资源不存在
  500: 服务器内部错误